+++ src/main/java/uk/q3c/krail/core/i18n/VaadinCurrentLocale.java
@@ -22,6 +22,7 @@ import org.slf4j.LoggerFactory;
 import uk.q3c.krail.core.eventbus.SessionBusProvider;
 import uk.q3c.krail.core.guice.uiscope.UIScoped;
 import uk.q3c.krail.core.guice.vsscope.VaadinSessionScoped;
+import uk.q3c.krail.core.option.VaadinOptionContext;
 import uk.q3c.krail.core.shiro.SubjectProvider;
 import uk.q3c.krail.core.ui.BrowserProvider;
 import uk.q3c.krail.core.user.status.UserStatusBusMessage;
@@ -29,7 +30,6 @@ import uk.q3c.krail.eventbus.BusMessage;
 import uk.q3c.krail.i18n.*;
 import uk.q3c.krail.i18n.bind.I18NModule;
 import uk.q3c.krail.option.Option;
-import uk.q3c.krail.option.OptionContext;
 import uk.q3c.krail.option.OptionKey;
 
 import java.util.Locale;
@@ -64,7 +64,7 @@ import java.util.Set;
  * @date 5 May 2014
  */
 
-public class VaadinCurrentLocale implements CurrentLocale, OptionContext {
+public class VaadinCurrentLocale implements CurrentLocale, VaadinOptionContext {
 
     public static final OptionKey<Locale> optionPreferredLocale = new OptionKey<>(Locale.UK, VaadinCurrentLocale.class, LabelKey.Preferred_Locale,
             DescriptionKey.Preferred_Locale);
@@ -202,7 +202,7 @@ public class VaadinCurrentLocale implements CurrentLocale, OptionContext {
 
 
     @Override
-    public Option getOption() {
+    public Option optionInstance() {
         return option;
     }