+++ src/main/java/uk/q3c/krail/core/option/VaadinOptionContext.java
@@ -1,0 +1,10 @@
+package uk.q3c.krail.core.option;
+
+import com.vaadin.data.Property;
+import uk.q3c.krail.option.OptionContext;
+
+/**
+ * Created by David Sowerby on 03 Aug 2017
+ */
+public interface VaadinOptionContext extends OptionContext<Property.ValueChangeEvent> {
+}