+++ src/main/java/uk/q3c/krail/core/view/component/DefaultUserNavigationMenu.java
@@ -25,16 +25,16 @@ import uk.q3c.krail.core.i18n.DescriptionKey;
 import uk.q3c.krail.core.i18n.LabelKey;
 import uk.q3c.krail.core.navigate.sitemap.UserSitemapLabelChangeMessage;
 import uk.q3c.krail.core.navigate.sitemap.UserSitemapStructureChangeMessage;
+import uk.q3c.krail.core.option.VaadinOptionContext;
 import uk.q3c.krail.core.vaadin.ID;
 import uk.q3c.krail.option.Option;
-import uk.q3c.krail.option.OptionContext;
 import uk.q3c.krail.option.OptionKey;
 
 import java.util.Optional;
 
 @Listener
 @SubscribeTo(SessionBus.class)
-public class DefaultUserNavigationMenu extends MenuBar implements OptionContext, UserNavigationMenu {
+public class DefaultUserNavigationMenu extends MenuBar implements VaadinOptionContext, UserNavigationMenu {
 
     protected static final OptionKey<Integer> optionKeyMaximumDepth = new OptionKey<>(10, DefaultUserNavigationMenu.class, LabelKey.Maxiumum_Depth,
             DescriptionKey.Maximum_Menu_Depth);
@@ -107,7 +107,7 @@ public class DefaultUserNavigationMenu extends MenuBar implements OptionContext,
 
 
     @Override
-    public Option getOption() {
+    public Option optionInstance() {
         return option;
     }