+++ src/main/java/uk/q3c/krail/core/view/component/LocaleContainer.java
@@ -22,9 +22,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.q3c.krail.core.i18n.DescriptionKey;
 import uk.q3c.krail.core.i18n.LabelKey;
+import uk.q3c.krail.core.option.VaadinOptionContext;
 import uk.q3c.krail.i18n.SupportedLocales;
 import uk.q3c.krail.option.Option;
-import uk.q3c.krail.option.OptionContext;
 import uk.q3c.krail.option.OptionKey;
 import uk.q3c.krail.util.ResourceUtils;
 
@@ -32,7 +32,7 @@ import java.io.File;
 import java.util.Locale;
 import java.util.Set;
 
-public class LocaleContainer extends IndexedContainer implements OptionContext {
+public class LocaleContainer extends IndexedContainer implements VaadinOptionContext {
 
 
     public enum PropertyName {
@@ -109,7 +109,7 @@ public class LocaleContainer extends IndexedContainer implements OptionContext {
 
 
     @Override
-    public Option getOption() {
+    public Option optionInstance() {
         return option;
     }