+++ src/main/java/uk/q3c/krail/i18n/persist/source/DefaultPatternSourceProvider.java
@@ -75,7 +75,7 @@ public class DefaultPatternSourceProvider implements PatternSourceProvider {
      */
 
     @Override
-    public Option getOption() {
+    public Option optionInstance() {
         return option;
     }