+++ src/test/groovy/uk/q3c/krail/core/persist/clazz/i18n/DefaultClassPatternDaoTest.groovy
@@ -208,7 +208,7 @@ class DefaultClassPatternDaoTest extends Specification {
 
     def "getOption()"() {
         expect:
-        dao.getOption() == option
+        dao.optionInstance() == option
     }
 
     def "set and get write file"() {